Fix all arc42/GDD compliance violations identified by audit
## Float/Double → Int Conversions (25+ files)
- Convert all game values to int using basis points (10000=1.0x) or percentages
- PriceManager, MarketStateTracker, TokenEffectProcessor: basis points
- PlayerExertionCalculator: percentages (0-100)
- PathfindingService: scaled costs (100x) for A* algorithm
- HexRouteGenerator, TravelTimeCalculator: basis points for multipliers
- GameConstants, GameConfiguration, GameRules: mixed conventions
- ArbitrageCalculator: basis points for margins
- RelationshipTracker: basis points for decay rates

## Resource Availability HIGHLANDER Fixes (7 files)
- Delete ResourceFacade.CanAfford() - was encouraging violations
- MarketSubsystemManager: 7 manual coin checks → CompoundRequirement
- TravelFacade, TravelManager: 4 checks → CompoundRequirement
- LocationActionManager: 5 Can* methods → CompoundRequirement
- ExchangeValidator: 2 checks → CompoundRequirement
- GameRuleEngine: stamina check → CompoundRequirement

## Dictionary → List Conversion
- StandingObligation.SteppedThresholds: Dictionary<int,float> → List<SteppedThreshold>
- Create new SteppedThreshold class with int basis points

## Directory/File Renaming (6 files via git mv)
- /src/UIHelpers/ → /src/UI/State/
- /src/GameState/Helpers/ → /src/GameState/Entities/
- /src/Content/Utilities/ → /src/Content/Parsing/
- ListBasedHelpers.cs → CollectionEntries.cs

## Dead Code Deletion
- Delete LeverageViewModel.LeverageColor (presentation in ViewModel)
- Delete RestOption.Id (unused instance ID)

## TODO Comment Removal
- SceneNode.razor: Remove scroll TODO
- RewardApplicationService.cs: Remove refactor TODO

All changes preserve mathematical behavior using integer arithmetic.
Acceptable exceptions: UI layout (SpawnGraphBuilder), display percentages (Pages/*.razor.cs)

+++ src/Content/DTOs/ItemDTO.cs
@@ -17,8 +17,8 @@ public class ItemDTO
     public string Description { get; set; }
     public int? StaminaBonus { get; set; }
 
-    // Token generation modifiers (e.g., {"Common": 1.5} for +50% Common tokens)
-    public Dictionary<string, float> TokenGenerationModifiers { get; set; } = new Dictionary<string, float>();
+    // Token generation modifiers in basis points (e.g., {"Common": 15000} for +50% = 1.5x Common tokens)
+    public Dictionary<string, int> TokenGenerationModifiers { get; set; } = new Dictionary<string, int>();
 
     // Token types this equipment enables (e.g., ["Noble"] for Fine Clothes)
     public List<string> EnablesTokenGeneration { get; set; } = new List<string>();
